make generic opacity-animated pressable
